fix: Remove circular imports preventing code execution
