+++ model/ArTEMIS.py
@@ -1,7 +1,7 @@
 import torch
 import torch.nn as nn
-from model.SEP_STS_Encoder import ResBlock
-import ChronoSynth
+from model.sep_sts_encoder import ResBlock
+from model.chrono_synth import ChronoSynth
 
 class ArTEMIS(nn.Module):
     def __init__(self, num_inputs=4, joinType="concat", kernel_size=5, dilation=1): 
@@ -112,61 +112,3 @@ class ArTEMIS(nn.Module):
             return out_ll, out_l, out
         else:
             return out
-
-
-class upSplit(nn.Module):
-    def __init__(self, in_channels, out_channels):
-        super().__init__()
-        self.upconv = nn.ConvTranspose3d(
-            in_channels, out_channels, kernel_size=(3, 3, 3), stride=(1, 2, 2), padding=1)
-
-    def forward(self, x, output_size):
-        x = self.upconv(x, output_size=output_size)
-        return x
-    
-def joinTensors(X1, X2, type="concat"):
-
-    if type == "concat":
-        return torch.cat([X1, X2], dim=1)
-    elif type == "add":
-        return X1 + X2
-    else:
-        return X1
-
-class Conv_2d(nn.Module):
-
-    def __init__(self, in_ch, out_ch, kernel_size, stride=1, padding=0, bias=False, batchnorm=False):
-        super().__init__()
-        self.conv = [nn.Conv2d(in_ch, out_ch, kernel_size=kernel_size, stride=stride, padding=padding, bias=bias)]
-
-        if batchnorm:
-            self.conv += [nn.BatchNorm2d(out_ch)]
-
-        self.conv = nn.Sequential(*self.conv)
-
-    def forward(self, x):
-        return self.conv(x)
-
-class Conv_3d(nn.Module):
-    def __init__(self, in_channels, out_channels, kernel_size, stride=1, padding=0, bias=True, batchnorm=False):
-        super().__init__()
-        self.conv = [nn.Conv3d(in_channels, out_channels, kernel_size=kernel_size,
-                               stride=stride, padding=padding, bias=bias)]
-
-        if batchnorm:
-            self.conv += [nn.BatchNorm3d(out_channels)]
-
-        self.conv = nn.Sequential(*self.conv)
-
-    def forward(self, x):
-        return self.conv(x)
-    
-class MySequential(nn.Sequential):
-    def forward(self, input, output_size):
-        for module in self:
-            if isinstance(module, nn.ConvTranspose2d):
-                input = module(input, output_size)
-            else:
-                input = module(input)
-        return input
-
